Make incompatible pointer types an error. (#9805)
  - For anything we build inside OMCompiler/, excluding OMCompiler/3rdParty,
    for the incompatible pointer types warning to error.

  - Fix unmatched buffer size in `gbode_util.c`

  - Rename the `clang-qt5` stage to `clang-qt5-omedit-testsuite`

+++ OMCompiler/SimulationRuntime/c/util/omc_msvc.c
@@ -313,7 +313,7 @@ int omc_dladdr(void *addr, Dl_info *info)
 {
   HANDLE hProcess;
   DWORD dwModuleBase;
-  DWORD displacement;
+  DWORD64 displacement;
   char sModuleName[MAX_PATH + 1];
   sModuleName[MAX_PATH] = '\0';
